refactor(mapper): restructure app
