+++ src/mapper/src/lib/components/editor/index.ts
@@ -1,0 +1,1 @@
+// Components related to the rich text editor
