+++ src/mapper/src/lib/components/editor/toolbar.svelte
