+++ src/mapper/src/lib/components/map/index.ts
@@ -1,0 +1,1 @@
+// Components related to the map display
