+++ src/mapper/src/lib/components/map/layer-switcher.svelte
