+++ src/mapper/src/lib/components/more/index.ts
@@ -1,0 +1,1 @@
+// Components for detailed view of task events
