+++ src/mapper/src/lib/components/more/skeleton/activities.svelte
