+++ src/mapper/src/lib/utils/qrcode.ts
