+++ src/mapper/src/routes/+layout.svelte
@@ -2,7 +2,7 @@
 	import '@hotosm/ui/dist/style.css';
 	import '@hotosm/ui/dist/hotosm-ui';
 	import { setBasePath } from '@hotosm/ui/dist/hotosm-ui';
-	import Toast from '$lib/components/common/toast.svelte';
+	import Toast from '$lib/components/toast.svelte';
 
 	setBasePath('/mapnow');
