+++ src/mapper/src/routes/+layout.ts
@@ -2,7 +2,7 @@ export const prerender = false;
 export const ssr = false;
 
 import 'virtual:uno.css';
-// import { initDb } from '$lib/init-db';
+// import { initDb } from '$lib/db/pglite';
 
 export function load() {
 	return {
